+++ CHANGELOG.md
@@ -3,15 +3,31 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Changelog](http://keepachangelog.com/).
 
+
 ## Unreleased
 
 ### New
-
 ### Changes
 
 ### Fixes
 
+## v2.2.0 (2022-10-24)
+
+### New
+- new --env-file CLI parameter on apply and destroy enabling specific declaration of the dotenv file to use
+
+### Changes
+- modified output messaging to use logger instead of bolded print when loading project configs
+
+### Fixes
+- exit code on on failed modules changed from 0 1o 1
+- changed typo `Deployemnt` to `Deployment` everywhere
+- improved validation and error handling when AccountId is not resolvable from EnvVariable
+- truncate and generate unique role and stack names when identifier is too long
+
+
 ## v2.1.0 (2022-10-03)
+
 
 ### New
 - added updated multi-account diagram source for docs